\name{filesByCodes}
\alias{filesByCodes}
\title{Relation between files and codes.}
\description{
Return a data frame which indicates what codes are associted with each file.
}
\usage{
filesByCodes(codingTable = c("coding", "coding2"))
}
\arguments{
  \item{codingTable}{ name of the coding table in rqda database.}
}
\details{}
\value{
A data frame
}
\author{Ronggui HUANG}
\seealso{}
\examples{
}