\name{getMemos}
\alias{getMemos}
\title{
Collection of code memos
}
\description{
This function collects all code memos into an object and displays them in a widget.
}
\usage{
getMemos(type = "codes")
}
\arguments{
  \item{type}{ Currently, only "codes" is supported.

}
}
%\details{}
\value{
  An object of class c("memos", "Info4Widget", "data.frame").
}
\author{
Ronggui HUANG
}
%\seealso{}
