\name{RQDA-internal}
\alias{addcode}
\alias{ClearMark}
\alias{close_proj}
\alias{codesupdate}
\alias{fnamesupdate}
\alias{HL}
\alias{importfile}
\alias{is_projOpen}
\alias{mark}
\alias{new_proj}
\alias{Proj_MemoButton}
\alias{File_MemoButton}
\alias{CodeMemoButton}
\alias{CodingMemoButton}
\alias{File_RenameButton}
%% to be documented independently.
\alias{setEncoding}
\alias{rename}
%% to be documented independently.
\alias{open_proj}
\alias{rqdameta}
\alias{retrieval}
\alias{sindex}
\alias{AddCase}
\alias{AddCaseButton}
\alias{AddCodeButton}
\alias{AddCodeCatButton}
\alias{AddTodbTable}
\alias{AddWebSearchButton}
\alias{CaseMark_Button}
\alias{CaseMemoButton}
\alias{CaseNamesUpdate}
\alias{CaseNamesWidgetMenu}
\alias{Case_RenameButton}
\alias{CloseProjectButton}
\alias{CodeCatAddToButton}
\alias{CodeCatDropFromButton}
\alias{CodeCat_RenameButton}
\alias{CodeNamesUpdate}
\alias{DeleteCaseButton}
\alias{DeleteCodeButton}
\alias{DeleteCodeCatButton}
\alias{DeleteFileButton}
\alias{ExtendButton}
\alias{FileNamesUpdate}
\alias{FreeCode_RenameButton}
\alias{HL_ALLButton}
\alias{Handler}
\alias{ImportFile}
\alias{ImportFileButton}
\alias{Mark_Button}
\alias{NewProjectButton}
\alias{OpenProjectButton}
\alias{ProjectInforButton}
\alias{RetrievalButton}
\alias{Unmark_Button}
\alias{UpdateCodeofCatWidget}
\alias{UpdateTableWidget}
\alias{ViewFileButton}
\alias{enc}
\alias{retrieval2}
\alias{FileCatAddToButton}
\alias{FileCatDropFromButton}
\alias{FileCat_RenameButton}
\alias{AddFileCatButton}
\alias{DeleteFileCatButton}
\alias{UpdateFileofCatWidget}
\alias{ViewFileFun}
\alias{AddFileToCaselinkage}
%\alias{AddFileToCaseMenu} % renamed as FileNamesWidgetMenu
\alias{HL_Case}
\alias{UpdateFileofCaseWidget}
\alias{MemoWidget}
\alias{addSettingGUI}
\alias{AddToFileCategory}
\alias{CaseUnMark_Button}
\alias{CodeCatWidgetMenu}
\alias{FileCatWidgetMenu}
\alias{FileofCaseWidgetMenu}
\alias{FileNamesWidgetMenu}
\alias{MarkCodeFun}
\alias{MarkCaseFun}
\alias{CodeofCatWidgetMenu}
\alias{CodesNamesWidgetMenu}
\alias{FileofCatWidgetMenu}
\alias{backup_proj}
\alias{BackupProjectButton}
\alias{UncodedFileNamesUpdate}
\alias{ProjectMemoWidget}
\alias{FileNameWidgetUpdate}
\alias{OrderByTime}
\alias{AutoCoding}
\alias{c2InfoFun}
\alias{CodingInfoButton}
\alias{gselect.list}
\alias{RunOnSelected}
%% add related alias functions here.

\title{Internal Functions}
\description{
  Internal RQDA functions.
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }