% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCases}
\alias{getCases}
\alias{getCaseIds}
\alias{getCaseNames}
\title{Get the Case ID and Case Name.}
\usage{
getCaseIds(fid = getFileIds(), nFiles = FALSE)
  
getCaseNames(caseId = getCaseIds(nFiles = FALSE))
  
getCases(fid, names = TRUE)
}
\arguments{
\item{fid}{numeric vector, the file IDs.}

\item{names}{logical.}

\item{nFiles}{logical, return the number of files that belong to a case.}

\item{caseId}{numeric vector, the case IDs.}
}
\value{
\code{getCaseIds} returns a data frame of two columns when nFiles is
TRUE, and a numeric vector when FALSE.

\code{getCaseNames} returns a character vector or NULL if no cases are
associated with the file IDs.

\code{getNames} returns the names of cases when names is TRUE, id of files
when FALSE.
}
\description{
Return cases IDs or names which a set of files belong to.
}
\details{
\code{getCaseIds} returns the case IDs which a file belongs to given the file
IDs.
\code{getCaseNames} returns the case Names given the case IDs.
\code{getCases} returns the case Names or IDs depending on the argument of
names. It is a wrapper of \code{getCaseIds} and \code{getCaseNames}.
}
\examples{
\dontrun{
getCaseNames(getCaseIds(getFileIds("filecategory")))
}
}
\seealso{
\code{\link{getFileIds}}
}
