% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gselect.list}
\alias{gselect.list}
\title{Select Items from a List}
\usage{
gselect.list(
  list,
  multiple = TRUE,
  title = NULL,
  height = getOption("widgetSize")[2],
  width = getOption("widgetSize")[1],
  ...
)
}
\arguments{
\item{list}{character vector. A list of items.}

\item{multiple}{logical: can more than one item be selected?}

\item{title}{optional character string for window title.}

\item{height}{integer. Height of the widget.}

\item{width}{integer. Width of the widget.}

\item{\dots}{Not used currently.}
}
\value{
A character vector of selected items with UTF-8 encoding. If no
item was selected and 'OK' is clicked, it returns length 0 character
vector. If 'Cancel' is clicked, '""' is returned.
}
\description{
Select item(s) from a character vector.
}
\details{
GTK version of \code{\link[utils]{select.list}}.
}
\note{
The license of this function is subject to interpretation of the first author.
}
\examples{
\dontrun{
select.list(sort(.packages(all.available = TRUE)))
}
}
\seealso{
\code{\link[utils]{select.list}}
}
