% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodesFun.R
\name{retrieval}
\alias{retrieval}
\title{Retrieval of codings conditional on the file id.}
\usage{
retrieval(Fid = NULL, order = c("fname", "ftime", "ctime"),
CodeNameWidget = .rqda$.codes_rqda,codingTable="coding")
}
\arguments{
\item{Fid}{Numeric vector, the file id.}

\item{order}{The method of sort of retrieved codings.}

\item{CodeNameWidget}{The name of code list widget.}

\item{codingTable}{name of sqlite data table where codings are stored.
It should be either "coding" or "coding2"}
}
\value{
A \code{\link[gWidgets2]{gtext}} widget is open and all the codings
are pushed into that widget.
}
\description{
To retrieve the codings of a selected code from specific set of files.
}
\details{
This function retrieves the codings of a selected code from
\code{CodeNameWidget}, given that all the codings are from a set of
files which are determined by \code{Fid}.
}
\seealso{
\code{\link{getFileIds}}
}
