\name{RQDA-internal}
\alias{addcode}
\alias{ClearMark}
\alias{close_proj}
\alias{codesupdate}
\alias{fnamesupdate}
\alias{HL}
\alias{importfile}
\alias{is_projOpen}
\alias{mark}
\alias{new_proj}
\alias{setEncoding}
\alias{open_proj}
%%\alias{RQDA}
%% RQDA() appears in RQDA-package.rd
\alias{rqdameta}
\alias{retrieval}
\alias{sindex}
%% add related alias functions here.
\title{Internal Grid Functions}
\description{
  Internal RQDA functions.
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }